check ParseInt

+++ src/main/java/com/app/camel/processor/project/DeleteByIdProject.java
@@ -2,6 +2,7 @@ package com.app.camel.processor.project;
 
 import com.app.camel.dao.ProjectRepository;
 import com.app.camel.dao.impl.ProjectRepositoryImpl;
+import com.app.camel.fuction.IsInt;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.apache.camel.component.restlet.RestletConstants;
@@ -13,25 +14,39 @@ public class DeleteByIdProject implements Processor {
 
     private final static Logger LOGGER = Logger.getLogger(DeleteAllProject.class);
     private final ProjectRepository projectRepository = new ProjectRepositoryImpl();
+    private final IsInt isInt= new IsInt();
 
     @Override
     public void process(Exchange exchange) throws Exception {
 
         String id = exchange.getIn().getHeader("id", String.class);
 
-        if(projectRepository.delete(Integer.parseInt(id)))
+
+        if(isInt.isInt(id))
         {
-            Response response = exchange.getIn().getHeader(RestletConstants.RESTLET_RESPONSE, Response.class);
-            response.setStatus(Status.SUCCESS_NO_CONTENT);
-            exchange.getOut().setBody(response);
-            LOGGER.info("Delete project by id: " + id + " success");
+            if(projectRepository.delete(Integer.parseInt(id)))
+            {
+                Response response = exchange.getIn().getHeader(RestletConstants.RESTLET_RESPONSE, Response.class);
+                response.setStatus(Status.SUCCESS_NO_CONTENT);
+                exchange.getOut().setBody(response);
+                LOGGER.info("Delete project by id: " + id + " success");
+            }
+            else
+            {
+                Response response = exchange.getIn().getHeader(RestletConstants.RESTLET_RESPONSE, Response.class);
+                response.setStatus(Status.REDIRECTION_NOT_MODIFIED);
+                exchange.getOut().setBody(response);
+                LOGGER.info("Delete project by id: " + id + " failed");
+            }
         }
         else
         {
             Response response = exchange.getIn().getHeader(RestletConstants.RESTLET_RESPONSE, Response.class);
-            response.setStatus(Status.REDIRECTION_NOT_MODIFIED);
+            response.setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
             exchange.getOut().setBody(response);
-            LOGGER.info("Delete project by id: " + id + " failed");
+            LOGGER.info("Wrong id");
         }
+
+
     }
 }
